+++ src/propertyFields/collectionData/ICustomCollectionField.ts
@@ -21,6 +21,22 @@ export interface ICustomCollectionField {
    * Dropdown options. Only nescessary when dropdown type is used.
    */
   options?: IDropdownOption[];
+  /**
+   * Input placeholder text.
+   */
+  placeholder?: string;
+  /**
+   * Default value for the field
+   */
+  defaultValue?: any;
+  /**
+   * The method is used to get the validation error message and determine whether the input value is valid or not.
+   *
+   * When it returns string:
+   * - If valid, it returns empty string.
+   * - If invalid, the field will show a red border
+   */
+  onGetErrorMessage?: (value: any) => string | Promise<string>;
 }
 
 export enum CustomCollectionFieldType {