+++ src/propertyFields/collectionData/collectionIconField/CollectionIconField.tsx
@@ -5,21 +5,16 @@ import { TextField } from 'office-ui-fabric-react/lib/components/TextField';
 import { Icon } from 'office-ui-fabric-react/lib/components/Icon';
 
 export class CollectionIconField extends React.Component<ICollectionIconFieldProps, {}> {
-  constructor(props: ICollectionIconFieldProps) {
-    super(props);
-
-    this.state = {
-      iconName: null
-    };
-  }
 
   public render(): React.ReactElement<ICollectionIconFieldProps> {
     return (
       <div className={styles.iconField}>
-        <TextField placeholder={this.props.field.title}
+        <TextField placeholder={this.props.field.placeholder || this.props.field.title}
+                   className={styles.collectionDataField}
                    value={this.props.item[this.props.field.id] ? this.props.item[this.props.field.id] : ""}
                    required={this.props.field.required}
-                   onChanged={(value) => this.props.fOnValueChange(this.props.field.id, value)} />
+                   onChanged={(value) => this.props.fOnValueChange(this.props.field.id, value)}
+                   onGetErrorMessage={(value) => this.props.fValidation(this.props.field, value)} />
         <Icon iconName={this.props.item[this.props.field.id] ? this.props.item[this.props.field.id] : ""} />
       </div>
     );