+++ src/propertyFields/collectionData/collectionNumberField/ICollectionNumberFieldState.ts
@@ -1,0 +1,3 @@
+export interface ICollectionNumberFieldState {
+  errorMessage: string;
+}